[SELC-5004] feat: Predisposition for reading the language from query parameter
