+++ package.json
@@ -11,7 +11,7 @@
     "@mui/x-data-grid": "^5.6.1",
     "@mui/x-data-grid-generator": "^5.6.1",
     "@pagopa/mui-italia": "^1.1.0",
-    "@pagopa/selfcare-common-frontend": "^1.34.16",
+    "@pagopa/selfcare-common-frontend": "^1.34.19",
     "@pagopa/ts-commons": "^10.2.1",
     "@reduxjs/toolkit": "^1.6.2",
     "@types/lodash": "^4.14.172",
